Write flexible init and refactor folder structure
